Project/testing/requirements.txt
pytest==6.2.0
pytest-xdist==2.2.1
selenium==3.141.0
allure-pytest==2.8.32
requests==2.25.1
Faker==8.0.0